Fix using a function for status field (PHNX-8691) (#21)
Motivation
--------

The status field on a mocked route is supposed to be able to be a function. This lets you return different status codes from the mocked route depending on the input. However, setting it to a function currently fails.

Modifications
---------

Updated check to allow status to be a function

Added tests

https://centeredge.atlassian.net/browse/PHNX-8691