Support more content types (PHNX-8794) (#22)
Motivation
---------

Currently, mocks only parse the body of a request if the request has a content type of "application/json". Some requests to Advantage have a content type of "application/json-patch+json". When mocking against those requests, the request body is not parsed.

Modfications
---------

Added support for more content types

https://centeredge.atlassian.net/browse/PHNX-8794